// frontend/src/apiService/models/AnswerPublicExtended.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SentimentAnalysisPublic } from './SentimentAnalysisPublic';
import {
    SentimentAnalysisPublicFromJSON,
    SentimentAnalysisPublicFromJSONTyped,
    SentimentAnalysisPublicToJSON,
    SentimentAnalysisPublicToJSONTyped,
} from './SentimentAnalysisPublic';

/**
 * 
 * @export
 * @interface AnswerPublicExtended
 */
export interface AnswerPublicExtended {
    /**
     * 
     * @type {string}
     * @memberof AnswerPublicExtended
     */
    questionId: string;
    /**
     * 
     * @type {string}
     * @memberof AnswerPublicExtended
     */
    text: string;
    /**
     * 
     * @type {string}
     * @memberof AnswerPublicExtended
     */
    id: string;
    /**
     * 
     * @type {SentimentAnalysisPublic}
     * @memberof AnswerPublicExtended
     */
    sentiment?: SentimentAnalysisPublic | null;
}

/**
 * Check if a given object implements the AnswerPublicExtended interface.
 */
export function instanceOfAnswerPublicExtended(value: object): value is AnswerPublicExtended {
    if (!('questionId' in value) || value['questionId'] === undefined) return false;
    if (!('text' in value) || value['text'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    return true;
}

export function AnswerPublicExtendedFromJSON(json: any): AnswerPublicExtended {
    return AnswerPublicExtendedFromJSONTyped(json, false);
}

export function AnswerPublicExtendedFromJSONTyped(json: any, ignoreDiscriminator: boolean): AnswerPublicExtended {
    if (json == null) {
        return json;
    }
    return {
        
        'questionId': json['question_id'],
        'text': json['text'],
        'id': json['id'],
        'sentiment': json['sentiment'] == null ? undefined : SentimentAnalysisPublicFromJSON(json['sentiment']),
    };
}

export function AnswerPublicExtendedToJSON(json: any): AnswerPublicExtended {
    return AnswerPublicExtendedToJSONTyped(json, false);
}

export function AnswerPublicExtendedToJSONTyped(value?: AnswerPublicExtended | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'question_id': value['questionId'],
        'text': value['text'],
        'id': value['id'],
        'sentiment': SentimentAnalysisPublicToJSON(value['sentiment']),
    };
}