// frontend/src/apiService/models/Answer.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Answer
 */
export interface Answer {
    /**
     * 
     * @type {string}
     * @memberof Answer
     */
    questionId: string;
    /**
     * 
     * @type {string}
     * @memberof Answer
     */
    text: string;
    /**
     * 
     * @type {string}
     * @memberof Answer
     */
    id?: string;
}

/**
 * Check if a given object implements the Answer interface.
 */
export function instanceOfAnswer(value: object): value is Answer {
    if (!('questionId' in value) || value['questionId'] === undefined) return false;
    if (!('text' in value) || value['text'] === undefined) return false;
    return true;
}

export function AnswerFromJSON(json: any): Answer {
    return AnswerFromJSONTyped(json, false);
}

export function AnswerFromJSONTyped(json: any, ignoreDiscriminator: boolean): Answer {
    if (json == null) {
        return json;
    }
    return {
        
        'questionId': json['question_id'],
        'text': json['text'],
        'id': json['id'] == null ? undefined : json['id'],
    };
}

export function AnswerToJSON(json: any): Answer {
    return AnswerToJSONTyped(json, false);
}

export function AnswerToJSONTyped(value?: Answer | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'question_id': value['questionId'],
        'text': value['text'],
        'id': value['id'],
    };
}